+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 =============================================================
 
 ### API breaking changes
+* Xcode6 is now supported from the main Xcode project Realm.xcodeproj. Xcode5 is no
+  longer supported. 
 
 ### Enhancements